Moved all rca utils to one file
